version 0.3

+++ lib/dao/cuentaDao.dart
@@ -1,3 +1,6 @@
+import 'dart:convert';
+import 'dart:io';
+
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:cuentas_android/dao/userDao.dart';
 import 'package:cuentas_android/models/Cuenta.dart';
@@ -5,40 +8,27 @@ import 'package:cuentas_android/models/Gasto.dart';
 import 'package:cuentas_android/models/Mes.dart';
 import 'package:cuentas_android/widgets/toast.dart';
 import 'package:firebase_auth/firebase_auth.dart';
+import 'package:flutter/services.dart';
 import 'package:get/get_rx/src/rx_types/rx_types.dart';
 
 class cuentaDao {
   CollectionReference ref = FirebaseFirestore.instance.collection('cuenta');
-  User? user = Auth().currentUser ?? null;
+  User? user = Auth().currentUser;
   static int count = 0;
 
-  Future obtenerDatos() async {
-    /*
-    final snapshot = await ref.get();
-    Values().cuentas = snapshot.docs.map((doc) => Cuenta.fromJson(doc.data() as Map<String, dynamic>)).toList();
-    */
+  Future<List<Cuenta>> getDatosJson() async {
+    String jsonCuentaWeb =
+        '{"id":"eqqpOnfDzsea4X6VEPRf2RnF6jc2","Nombre":"App test","Meses":[{"Gastos":[{"nombre":"Casa","valor":850.0,"anno":2024,"mes":9,"dia":10,"tag":"indispensable"},{"nombre":"Bizum amigo","valor":-3.0,"anno":2024,"mes":9,"dia":10,"tag":""}],"Ingreso":1340.0,"Extras":[{"nombre":"Cine","valor":12.5,"anno":2024,"mes":9,"dia":4,"tag":"ocio"},{"nombre":"Compras del mes","valor":1500.0,"anno":2024,"mes":9,"dia":10,"tag":"compras"}],"NMes":"Septiembre","Anno":2024}],"posicion":3,"fijos":[{"nombre":"Casa","valor":850.0,"anno":2024,"mes":9,"dia":10,"tag":"indispensable"}],"deudas":[{"nombre":"Prestamo coche","valor":30678.0,"anno":2024,"mes":9,"dia":10,"tag":""}],"color":"#ffeb3f84","tags":["indispensable","","compras","ocio"]}';
 
-    /*if(Values().cuentas.isEmpty){
-      for(int i = 1;i<3;i++){
-        Values().cuentas.add( Cuenta(
-          id: i,
-          Meses: [],
-          Nombre: "nombre$i"
-        ));
-      }
-    }*/
+    Cuenta cuentaWeb = Cuenta.fromJson(await json.decode(jsonCuentaWeb));
+    return [cuentaWeb];
   }
 
-  Future<List<Cuenta>> getDatos() async {
-    /*return List.generate(2, (index) => Cuenta(
-      id: math.Random().nextDouble().toString(),
-      Meses: [],
-      Nombre: "nombre${index+1}",
-      posicion: index+1
-    ));
-    */
+  Future<List<Cuenta>> getDatos(bool isWeb) async {
+    if (isWeb) {
+      return await getDatosJson();
+    }
 
-    print(user);
     final snapshot = await ref.where('id', isEqualTo: user!.uid).get();
     List<Cuenta> ret = snapshot.docs
         .map((doc) => Cuenta.fromJson(doc.data() as Map<String, dynamic>))
@@ -47,25 +37,8 @@ class cuentaDao {
     return ret;
   }
 
-  Future migrardatos() async {
-    var snapshot = await ref.get();
-    List<Cuenta> datos = snapshot.docs
-        .map((doc) => Cuenta.fromJson(doc.data() as Map<String, dynamic>))
-        .toList();
-    int contador = 1;
-
-    for (Cuenta c in datos) {
-      c.id.value = user!.uid;
-      c.posicion.value = contador;
-
-      String doc = "${c.id}-${c.posicion}";
-      ref.doc(doc).set(c.toJson());
-      contador++;
-    }
-  }
-
   Future<Cuenta> crearNuevaCuenta(
-      String nombre, int posicion, String? color) async {
+      String nombre, int posicion, String? color, bool isWeb) async {
     String documento = "${user!.uid}-$posicion";
     Cuenta c = Cuenta(
         Meses: RxList<Mes>(),
@@ -77,28 +50,37 @@ class cuentaDao {
         fijos: RxList<Gasto>(),
         tags: RxList<String>());
 
-    await ref.doc(documento).set(c.toJson());
+    if (isWeb) {
+      await ref.doc(documento).set(c.toJson());
+    }
 
     return c;
   }
 
-  Future almacenarDatos(Cuenta c) async {
-    await ref.doc("${c.id}-${c.posicion}").update(c.toJson());
-    
-      showToast(text: 'guardado correctamente');
+  Future almacenarDatos(Cuenta c, bool isWeb) async {
+    if (!isWeb) {
+      await ref.doc("${c.id}-${c.posicion}").update(c.toJson());
+    }
+
+    showToast(text: 'guardado correctamente');
   }
 
-  Future deleteCuenta(Cuenta c) async {
-    await ref.doc("${c.id}-${c.posicion}").delete();
+  Future deleteCuenta(Cuenta c, bool isWeb) async {
+    if (!isWeb) {
+      await ref.doc("${c.id}-${c.posicion}").delete();
+    }
   }
 
-  Future<Cuenta> importFromJson(Map<String, dynamic> json, int posicion) async {
+  Future<Cuenta> importFromJson(
+      Map<String, dynamic> json, int posicion, bool isWeb) async {
     String documento = "${user!.uid}-$posicion";
     Cuenta c = Cuenta.fromJson(json);
     c.id.value = user!.uid;
     c.posicion.value = posicion;
 
-    await ref.doc(documento).set(c.toJson());
+    if (!isWeb) {
+      await ref.doc(documento).set(c.toJson());
+    }
 
     return c;
   }
